feat: centralizing InRoom UI types
